split vrt script separately
